EE-660: Updated Jira issue regex to accept issue id anywhere within branch name or commit message
Currently, the regex looks for the identifier at the start of the string (while also capturing all matches):

/^[a-zA-Z]+(?<!id)-\d+/g

This will accept branches like BLAH-123 and commits like BLAH-123: did some stuff, but will not accept feature/BLAH-123 or feat(BLAH-123): did some stuff.

We can update it to /[a-zA-Z]+(?<!id)-\d+/ to allow for those types.